[*] Fixed building for windows.

+++ modules/type/dag/dap_chain_cs_dag.c
@@ -1927,8 +1927,8 @@ static int s_cli_dag(int argc, char ** argv, void **a_str_reply)
                         size_t l_objs_count = 0;
                         l_objs = dap_global_db_get_all_sync(l_gdb_group_events,&l_objs_count);
                         char *ptr;
-                        uint l_limit = l_limit_str ? strtoull(l_limit_str, &ptr, 10) : 0;
-                        uint l_offset = l_offset_str ? strtoull(l_offset_str, &ptr, 10) : 0;
+                        size_t l_limit = l_limit_str ? strtoull(l_limit_str, &ptr, 10) : 0;
+                        size_t l_offset = l_offset_str ? strtoull(l_offset_str, &ptr, 10) : 0;
                         size_t l_arr_start = 0;
                         if (l_offset) {
                             l_arr_start = l_offset * l_limit;
